+++ firebase-perf/CHANGELOG.md
@@ -12,6 +12,7 @@
   now deprecated. As early as April 2024, we'll no longer release KTX modules. For details, see the
   [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
 
+* [changed] Ensure that AppStart trace contains a valid sessionID.
 
 ## Kotlin
 The Kotlin extensions library transitively includes the updated